requirements.txt
# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Document Generation
python-pptx>=0.6.23
python-docx>=1.1.0
markdown>=3.5

# Configuration & Templates
pyyaml>=6.0
jinja2>=3.1.0

# LLM Integration
requests>=2.31.0
anthropic>=0.72.0
google-generativeai>=0.3.0
openai>=1.0.0

# CLI & UI
rich>=13.0.0
typer>=0.9.0

# Data Processing
pydantic>=2.5.0
bibtexparser>=1.4.0
numpy>=1.24.0
tqdm>=4.66.0

# Vector Database & Embeddings
chromadb>=0.4.22

# Optional: Advanced PDF features
lxml>=5.0.0
Pillow>=10.0.0
